# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API preliminar del Marketplace",
        "description": "Documentación generada con L5-Swagger para la API del Marketplace Orientado a la ciudad de Chivilcoy.",
        "contact": {
            "name": "Santiago Castellano",
            "email": "scastellano10@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/track-contact": {
            "post": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Registrar un contacto con un negocio o producto",
                "description": "Registra cuando un usuario realiza una acción que indica interés en un negocio o producto (ej: ver teléfono, hacer clic en sitio web, etc.)",
                "operationId": "7d1ed8cef9f59a54e1af72f58aab2668",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Contact"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contacto registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Contacto registrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Negocio no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Negocio no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/businesses/{business}/contact-stats": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Obtener estadísticas de contactos para un negocio",
                "description": "Devuelve estadísticas de contactos para un negocio específico en los últimos 30 días",
                "operationId": "7826a17d4e175cf03d278e57edfd9002",
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "description": "ID del negocio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estadísticas de contactos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "last_30_days": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2025-10-26"
                                                    },
                                                    "count": {
                                                        "type": "integer",
                                                        "example": 5
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "by_type": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "contact_type": {
                                                        "type": "string",
                                                        "example": "phone_view"
                                                    },
                                                    "count": {
                                                        "type": "integer",
                                                        "example": 3
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 42
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Negocio no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Negocio no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Contact": {
                "required": [
                    "contactable_type",
                    "contactable_id",
                    "contact_type"
                ],
                "properties": {
                    "user_id": {
                        "description": "ID del usuario (nullable si no está autenticado)",
                        "type": "integer",
                        "example": 1
                    },
                    "contactable_type": {
                        "description": "Tipo de entidad contactada",
                        "type": "string",
                        "enum": [
                            "business",
                            "product"
                        ],
                        "example": "business"
                    },
                    "contactable_id": {
                        "description": "ID de la entidad contactada",
                        "type": "integer",
                        "example": 21
                    },
                    "contact_type": {
                        "description": "Tipo de contacto",
                        "type": "string",
                        "enum": [
                            "phone_view",
                            "email_view",
                            "address_view",
                            "website_click",
                            "social_click",
                            "copy_contact",
                            "favorite",
                            "share",
                            "map_view",
                            "catalog_download"
                        ],
                        "example": "phone_view"
                    },
                    "ip": {
                        "description": "IP del visitante",
                        "type": "string",
                        "example": "192.168.1.1"
                    },
                    "user_agent": {
                        "description": "User agent del navegador",
                        "type": "string",
                        "example": "Mozilla/5.0"
                    },
                    "referer": {
                        "description": "URL de referencia",
                        "type": "string",
                        "example": "https://chivimarket.arumasoft.com"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-26T12:34:56.000000Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}